fix(codegen): widen KOpt.axis u8 -> u32 (axis_ids exceed 255)
KOpt.axis was u8, but thvm axis_ids are GLOBAL -- they accumulate
across the whole DAG rather than being renumbered per-kernel -- so a
deep backward graph (the full 4-conv beautiful_mnist LeNet) reaches
axis_ids > 255.  Every KOpt construction cast the axis_id to u8
(`(u8)axc.axis_id[...]`), silently truncating e.g. 344 -> 88.

uop_dag_apply_split then called find_range(root, 88), found no RANGE
leaf with id 88, and returned 0.  kernel_apply_opt counted that as a
failure, so hand_opts applied NOTHING to those kernels (n_applied=0,
no LOCAL/UPCAST).  render_uop emitted them fully serial: a single
thread looping the entire output (e.g. 128x32x24x24 = 2.36M
iterations) with deeply-nested index expressions.  nvrtc's optimizer
chokes on that shape -- the conv1-forward recompute (kid 39) compiled
for >80 minutes at BS=128, which read as a "warmup hang".  Early
kernels (axis_ids < 256) were unaffected, which is why a 2-conv
half-LeNet worked but the full 4-conv net didn't.

Widen KOpt.axis to u32 and drop the truncating (u8) casts at every
construction site (hand_opts MV/GROUPTOP/UPCAST/UNROLL/LOCAL,
autotune seq parse).  Widen the two ctypes/FFI boundaries that passed
axis as a narrow int: py_kernel_apply_opt / py_uop_dag_apply_kopt
(uint8_t -> uint32_t, with matching argtypes in thvm.py) and the WL
thvm_wl_kernel_apply_opt (drop its (u8)axis cast; the mint arg was
already wide).  autotune's `axis > 255` reject is relaxed to u32.

After the fix kid 25 and kid 39 go from fully-serial (n_applied=0) to
parallelized (L=3 LOCAL + 2 UPCAST).  A second, independent large-
axis-id bug remains in render_uop's grid/local assignment (kid 39
still renders its batch axis as a serial loop and its LOCAL axes as
raw `tt`), filed as follow-up.  No regression on the 2-conv path
(test_apply_opt_dag 133/134 -- the one failure is a pre-existing
reduce-axis-shift bug unrelated to this change).

Includes THVM_HANDOPT_TRACE=1: per-kernel axis-type + n_applied
summary that localized this bug.

+++ py/csource/thvm_py.c
@@ -563,7 +563,7 @@ EXPORT void py_propose_tc_counters_reset(void) {
 // alongside the propose surface to compose the tinygrad-style BEAM
 // autotune loop in Python.
 EXPORT uint64_t py_kernel_apply_opt(uint32_t kid, uint8_t op,
-                                    uint8_t axis, uint32_t arg) {
+                                    uint32_t axis, uint32_t arg) {
   if (kid == 0 || kid >= KERNELS_NEXT) return 0;
   KernelEntry *ke = &KERNELS[kid];
   KOpt opt = { op, axis, arg };
@@ -576,7 +576,7 @@ EXPORT uint64_t py_kernel_apply_opt(uint32_t kid, uint8_t op,
 // for agents that prefer to manage the DAG directly without the
 // kid-keyed KERNELS[] machinery.
 EXPORT uint64_t py_uop_dag_apply_kopt(uint64_t root, uint8_t op,
-                                      uint8_t axis, uint32_t arg) {
+                                      uint32_t axis, uint32_t arg) {
   KOpt opt = { op, axis, arg };
   return uop_dag_apply_kopt(root, opt);
 }
